fix: Conflict resolutions
Move the mention utils into a web module and remote the now-unused subscribe
