+++ apps/web/src/lib/components/chat/Mention.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import { isMentionMatch, type MentionMatch } from '$lib/chat/utils';
+	import { isMentionMatch, type MentionMatch } from '$lib/chat/mentions';
 	import Self from '$lib/components/chat/Mention.svelte';
 
 	interface Props {
